components/query_engine/requirements.txt
chromadb==0.4.7
google-cloud-aiplatform==1.31.1
# google-cloud-aiplatform==1.31.0
langchain==0.0.276
# langchain==0.0.272
google-cloud-discoveryengine==0.10.0
google-generativeai==0.1.0
# --------- test
numpy==1.25.2
